backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psutil==6.1.1
watchdog==6.0.0
aiosqlite==0.20.0
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
